backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
pydantic==2.5.0
requests==2.31.0
pymongo==4.16.0
python-dotenv==1.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
google-api-python-client==2.187.0
python-multipart==0.0.6

